+++ pkg/util/tests/kubernetes.go
@@ -36,8 +36,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/client-go/kubernetes"
 
-	"github.com/arangodb/kube-arangodb/pkg/apis/analytics"
-	analyticsApi "github.com/arangodb/kube-arangodb/pkg/apis/analytics/v1alpha1"
 	"github.com/arangodb/kube-arangodb/pkg/apis/backup"
 	backupApi "github.com/arangodb/kube-arangodb/pkg/apis/backup/v1"
 	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
@@ -254,12 +252,6 @@ func CreateObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientSe
 			vl := *v
 			_, err := arango.SchedulerV1beta1().ArangoSchedulerCronJobs(vl.GetNamespace()).Create(context.Background(), vl, meta.CreateOptions{})
 			require.NoError(t, err)
-		case **analyticsApi.GraphAnalyticsEngine:
-			require.NotNil(t, v)
-
-			vl := *v
-			_, err := arango.AnalyticsV1alpha1().GraphAnalyticsEngines(vl.GetNamespace()).Create(context.Background(), vl, meta.CreateOptions{})
-			require.NoError(t, err)
 		case **networkingApi.ArangoRoute:
 			require.NotNil(t, v)
 
@@ -475,12 +467,6 @@ func UpdateObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientSe
 			vl := *v
 			_, err := arango.SchedulerV1beta1().ArangoSchedulerCronJobs(vl.GetNamespace()).Update(context.Background(), vl, meta.UpdateOptions{})
 			require.NoError(t, err)
-		case **analyticsApi.GraphAnalyticsEngine:
-			require.NotNil(t, v)
-
-			vl := *v
-			_, err := arango.AnalyticsV1alpha1().GraphAnalyticsEngines(vl.GetNamespace()).Update(context.Background(), vl, meta.UpdateOptions{})
-			require.NoError(t, err)
 		case **networkingApi.ArangoRoute:
 			require.NotNil(t, v)
 
@@ -629,12 +615,6 @@ func UpdateStatusObjects(t *testing.T, k8s kubernetes.Interface, arango arangoCl
 
 			vl := *v
 			_, err := arango.SchedulerV1beta1().ArangoSchedulerCronJobs(vl.GetNamespace()).UpdateStatus(context.Background(), vl, meta.UpdateOptions{})
-			require.NoError(t, err)
-		case **analyticsApi.GraphAnalyticsEngine:
-			require.NotNil(t, v)
-
-			vl := *v
-			_, err := arango.AnalyticsV1alpha1().GraphAnalyticsEngines(vl.GetNamespace()).UpdateStatus(context.Background(), vl, meta.UpdateOptions{})
 			require.NoError(t, err)
 		case **networkingApi.ArangoRoute:
 			require.NotNil(t, v)
@@ -799,11 +779,6 @@ func DeleteObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientSe
 
 			vl := *v
 			require.NoError(t, arango.SchedulerV1beta1().ArangoSchedulerCronJobs(vl.GetNamespace()).Delete(context.Background(), vl.GetName(), meta.DeleteOptions{}))
-		case **analyticsApi.GraphAnalyticsEngine:
-			require.NotNil(t, v)
-
-			vl := *v
-			require.NoError(t, arango.AnalyticsV1alpha1().GraphAnalyticsEngines(vl.GetNamespace()).Delete(context.Background(), vl.GetName(), meta.DeleteOptions{}))
 		case **networkingApi.ArangoRoute:
 			require.NotNil(t, v)
 
@@ -1220,21 +1195,6 @@ func RefreshObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientS
 			} else {
 				*v = vn
 			}
-		case **analyticsApi.GraphAnalyticsEngine:
-			require.NotNil(t, v)
-
-			vl := *v
-
-			vn, err := arango.AnalyticsV1alpha1().GraphAnalyticsEngines(vl.GetNamespace()).Get(context.Background(), vl.GetName(), meta.GetOptions{})
-			if err != nil {
-				if kerrors.IsNotFound(err) {
-					*v = nil
-				} else {
-					require.NoError(t, err)
-				}
-			} else {
-				*v = vn
-			}
 		case **networkingApi.ArangoRoute:
 			require.NotNil(t, v)
 
@@ -1518,14 +1478,6 @@ func SetMetaBasedOnType(t *testing.T, object meta.Object) {
 		v.SetSelfLink(fmt.Sprintf("/api/%s/%s/%s/%s",
 			schedulerApi.SchemeGroupVersion.String(),
 			scheduler.CronJobResourcePlural,
-			object.GetNamespace(),
-			object.GetName()))
-	case *analyticsApi.GraphAnalyticsEngine:
-		v.Kind = analytics.GraphAnalyticsEngineResourceKind
-		v.APIVersion = analyticsApi.SchemeGroupVersion.String()
-		v.SetSelfLink(fmt.Sprintf("/api/%s/%s/%s/%s",
-			analyticsApi.SchemeGroupVersion.String(),
-			analytics.GraphAnalyticsEngineResourcePlural,
 			object.GetNamespace(),
 			object.GetName()))
 	case *networkingApi.ArangoRoute: